+++ backend/.sequelizerc
@@ -1,6 +1,6 @@
 module.exports = {
-  config: 'src/config/config.json',
-  'seeders-path': 'src/seeders',
-  'migrations-path': 'src/migrations',
+  config: 'config/config.json',
+  'seeders-path': 'seeders',
+  'migrations-path': 'migrations',
   'models-path': 'src/models',
 };
