+++ backend/src/tag/tag.model.ts
@@ -1,4 +1,4 @@
-import { IFields } from 'models/types';
+import { IFields } from 'shared/database/types';
 import { Model } from 'sequelize';
 
 class Tag extends Model {
